Enemy, Animatiuons, Materials & Textures
A new Folder was created with the Animations. The Materials can be found in the "Enemy" Folder which is located in the Materials Folder. Finally the Meshes are in the Enemy Meshes Folder which can be found in the Meshes Folder. Everything is in their First Form and can be iterated over time.